remove all Graph.client Dependency from edge.scala
